get account dropdown working
