+++ client/src/app/dashboard/welcome-card.tsx
@@ -1,34 +1,15 @@
 import { AuthContext } from '@/components/auth-provider';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { firebaseAuth } from '@/lib/firebase';
-import { useQueryClient } from '@tanstack/react-query';
-import { signOut } from 'firebase/auth';
+import { Card, CardHeader, CardTitle } from '@/components/ui/card';
 import { useContext } from 'react';
 
 const WelcomeCard = (): JSX.Element => {
   const { currentUserState } = useContext<any>(AuthContext);
-  const queryClient = useQueryClient();
-
-  const Logout = (): void => {
-    signOut(firebaseAuth)
-      .then(() => {
-        queryClient.removeQueries();
-      })
-      .catch((err) => {
-        // TODO: Make this an alert dialog
-        console.log(err);
-      });
-  };
 
   return (
     <Card className="w-full">
       <CardHeader>
         <CardTitle>Hello, {currentUserState?.email}.</CardTitle>
       </CardHeader>
-      <CardContent>
-        <Button onClick={Logout}>Logout</Button>
-      </CardContent>
     </Card>
   );
 };
